The Prince by Niccolò Machiavelli

The Prince by Niccolò Machiavelli

Usage

ex_prince

Format

a data frame with The Prince book from Gutenberg Project separated by chapters

Source

<https://www.gutenberg.org/ebooks/1232>

Examples

str(ex_prince)
#> tibble [29 × 2] (S3: tbl_df/tbl/data.frame)
#>  $ title: chr [1:29] "DEDICATION" "THE PRINCE" "CHAPTER II." "CHAPTER III." ...
#>  $ text : chr [1:29] "To the Magnificent Lorenzo Di Piero De' Medici\n\n\nThose who strive to obtain the good graces of a prince are "| __truncated__ "CHAPTER I.\nHOW MANY KINDS OF PRINCIPALITIES THERE ARE, AND BY WHAT MEANS THEY ARE ACQUIRED\n\n\nAll states, al"| __truncated__ "CONCERNING HEREDITARY PRINCIPALITIES\n\n\nI will leave out all discussion on republics, inasmuch as in another "| __truncated__ "CONCERNING MIXED PRINCIPALITIES\n\n\nBut the difficulties occur in a new principality. And firstly, if it be\nn"| __truncated__ ...
ex_prince
#> # A tibble: 29 × 2
#>    title         text                                                           
#>    <chr>         <chr>                                                          
#>  1 DEDICATION    "To the Magnificent Lorenzo Di Piero De' Medici\n\n\nThose who…
#>  2 THE PRINCE    "CHAPTER I.\nHOW MANY KINDS OF PRINCIPALITIES THERE ARE, AND B…
#>  3 CHAPTER II.   "CONCERNING HEREDITARY PRINCIPALITIES\n\n\nI will leave out al…
#>  4 CHAPTER III.  "CONCERNING MIXED PRINCIPALITIES\n\n\nBut the difficulties occ…
#>  5 CHAPTER IV.   "WHY THE KINGDOM OF DARIUS, CONQUERED BY ALEXANDER, DID NOT RE…
#>  6 CHAPTER V.    "CONCERNING THE WAY TO GOVERN CITIES OR PRINCIPALITIES WHICH L…
#>  7 CHAPTER VI.   "CONCERNING NEW PRINCIPALITIES WHICH ARE ACQUIRED BY ONE'S OWN…
#>  8 CHAPTER VII.  "CONCERNING NEW PRINCIPALITIES WHICH ARE ACQUIRED EITHER BY TH…
#>  9 CHAPTER VIII. "CONCERNING THOSE WHO HAVE OBTAINED A PRINCIPALITY BY WICKEDNE…
#> 10 CHAPTER IX.   "CONCERNING A CIVIL PRINCIPALITY\n\n\nBut coming to the other …
#> # ℹ 19 more rows