Filter a graph / create an ego graph

filter a graph / create an ego graph by term and by the number of its neighbors

Usage

filter_ego(edges, nodes = NULL, filter_by, n_neighbours = 1)

filter_ego(edges, nodes = NULL, filter_by, n_neighbours = 1)

Arguments

edges

an edge dataframe

nodes

a node dataframe

filter_by

a term to filter the ego graph

n_neighbours

the number of neighbors

as_tbl

if TRUE, return a tbl_graph, if FALSE, return an igraph object

Examples

# creating sample data
nodes <- data.frame(id = 1:5, name = LETTERS[1:5])
edges <- data.frame(from = c(1, 1, 2, 3, 4, 1, 6, 7), to = c(2, 3, 4, 5, 5, 4, 7, 5))
filter_ego(edges, nodes, filter_by = 1, n_neighbours = 1)
#> Error in igraph::graph_from_data_frame(edges, directed = TRUE, vertices = nodes): Some vertex names in edge list are not listed in vertex data frame
filter_ego(edges, nodes, filter_by = 1, n_neighbours = 2)
#> Error in igraph::graph_from_data_frame(edges, directed = TRUE, vertices = nodes): Some vertex names in edge list are not listed in vertex data frame
# creating sample data
nodes <- data.frame(id = 1:5, name = LETTERS[1:5])
edges <- data.frame(from = c(1, 1, 2, 3, 4, 1, 6, 7), to = c(2, 3, 4, 5, 5, 4, 7, 5))
filter_ego(edges, nodes, filter_by = 1, n_neighbours = 1)
#> Error in igraph::graph_from_data_frame(edges, directed = TRUE, vertices = nodes): Some vertex names in edge list are not listed in vertex data frame
filter_ego(edges, nodes, filter_by = "jojo", n_neighbours = 2)
#> Error in igraph::graph_from_data_frame(edges, directed = TRUE, vertices = nodes): Some vertex names in edge list are not listed in vertex data frame