graph from text and query

from text and a query, generates the graph dataframe of co-ocurrence frequency

Usage

get_graph_from_txt(text, query, i_c = TRUE, by_sentence = TRUE, loop = FALSE)

Arguments

text

a vector of texts

query

query to filter the text

by_sentence

tokenize by sentence (default TRUE). If FALSE, tokenize by paragraph

loop

if TRUE (default: FALSE), returns a loop (self reference) in the graph

ic

ignore case (default TRUE)

Examples

# loading data
data(package = "txtnet")
# sample
txt_wiki[2:3]
#> [1] " "                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [2] "Brian Thompson, the 50-year-old CEO of the American health insurance company UnitedHealthcare, was shot and killed in Midtown Manhattan, New York City, on December 4, 2024. The shooting occurred early in the morning outside an entrance to the New York Hilton Midtown hotel.[4] Thompson was in the city to attend an annual investors' meeting for UnitedHealth Group, the parent company of UnitedHealthcare. Prior to his death, he faced criticism for the company's rejection of insurance claims, and his family reported that he had received death threats in the past. The suspect, initially described as a white man wearing a mask, fled the scene.[1] On December 9, 2024, authorities arrested 26-year-old Luigi Mangione in Altoona, Pennsylvania, and charged him with Thompson's murder in a Manhattan court.[5][6][7]"
txt_wiki |> get_graph_from_txt("York")
#> Error in get_pairs(parsePOS(filter_by_query(text, query = query, i_c = i_c,     by_sentence = by_sentence)), loop = loop): could not find function "get_pairs"
txt_wiki |> get_graph_from_txt("Police")
#> Error in get_pairs(parsePOS(filter_by_query(text, query = query, i_c = i_c,     by_sentence = by_sentence)), loop = loop): could not find function "get_pairs"