A Text Sample

A sample text from text from https://www.bbc.com/news/articles/c7ve36zg0e5o

Usage

text_sample

Format

An object of class character of length 1.

Examples

data(package = "txtnet")
head(text_sample)
#> [1] "A man is being questioned about the fatal shooting of a healthcare insurance boss in New York last week, officials have told the BBC's US partner CBS News.\nUnitedHealthcare boss Brian Thompson, 50, was fatally shot in the back on Wednesday morning last week outside the Hilton hotel in Midtown Manhattan.\nPolice say Thompson was targeted in a pre-planned killing, for which they do not yet have a motive. Nor have officers revealed a name of the suspect.\nIt is not clear if the person being questioned is the suspect who authorities have been searching for.\nInvestigators have been using surveillance photos, bullet casings with cryptic messages written on them, and the suspect's movements to track him down. They are also working with the FBI and authorities in other states...\nThe manhunt for a suspect who gunned down a healthcare chief executive in New York is now in its third day, with police chasing several different leads.\nUnitedHealthcare boss Brian Thompson, 50, was fatally shot in the back on Wednesday morning outside the Hilton hotel in Midtown Manhattan.\nPolice say Thompson was targeted in a pre-planned killing, for which they do not yet have a motive.\nInvestigators are using surveillance photos, bullet casings with cryptic messages written on them, and the suspect's movements to track him down. They are also working with the FBI and authorities in other states as the search expands beyond New York."